refactor(omegaform): extract submit + handleSubmitEffect
